[FL-1477] Archive improvements (#541)
* various small issues fixes
* proper fadding to favourites
* cleanups
* fix infrared extension
* allow subdirs in app tabs
* fix elements_scrollbar args
* bugfixes, looping list, halfbaked pinning
* pinning
* ui updates

+++ applications/gui/elements.c
@@ -67,7 +67,7 @@ void elements_scrollbar_pos(
     }
 }
 
-void elements_scrollbar(Canvas* canvas, uint8_t pos, uint8_t total) {
+void elements_scrollbar(Canvas* canvas, uint16_t pos, uint16_t total) {
     furi_assert(canvas);
 
     uint8_t width = canvas_width(canvas);
